[Fix] Corrigindo erros de outros apps para executar localmente (#299)
* [Fix] Corrigir nome do aplicativo para incluir o caminho correto

* [Fix] remover importações administrativas não utilizadas das configurações de URL

* [Fix] ajustar importação de modelos e remover classes de administrador não utilizadas

* [Fix] comentar classes e métodos não utilizados no modelo Motorista (duplicatas)

* [Fix] resolver conflitos de mesclagem no arquivo de migração inicial

* [Feature] adicionar estrutura inicial do aplicativo de esportes com rotas, views e templates

* Update apps/esportemuz/viewsets.py


* Update apps/esportemuz/viewsets.py


* [Feature] add championship management features, including closing championships and tracking match statistics

* Add database engine configuration and update allowed hosts in settings.py

---------
