+++ .env_sample
@@ -1,4 +1,5 @@
 SECRET_KEY=
+DB_ENGINE=
 DB_NAME=
 DB_USER=
 DB_PASSWORD=
