+++ apps/contratamuz/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 class ContratamuzConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'contratamuz'
+    name = 'apps.contratamuz'
     verbose_name = 'ContrataMuz'
     verbose_name_plural = 'ContrataMuz'
