+++ apps/contratamuz/urls.py
@@ -1,10 +1,7 @@
-from django.contrib import admin
-from django.urls import path
 from apps.contratamuz import views
+from django.urls import path
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
     path('', views.inicial, name='inicial'),
     path('vagas/', views.listar_vagas, name='listar_vagas'),
-    # adicione outras rotas conforme necessário
 ]
