+++ apps/covamuz/admin.py
@@ -1,6 +1,5 @@
-from .models import *
+from apps.covamuz.models import *
 from django.contrib import admin
-from .models import *
 
 admin.site.register(Pessoa)
 admin.site.register(Cemiterio)
@@ -9,4 +8,3 @@ admin.site.register(Tumulo)
 admin.site.register(HorarioVisitacao)
 admin.site.register(Hora)
 admin.site.register(Dia)
-admin.site.register(Pessoa)
