+++ apps/esportemuz/apps.py
@@ -5,6 +5,3 @@ class EsportemuzConfig(AppConfig):
     name = 'apps.esportemuz'
     verbose_name = 'EsporteMuz'
     verbose_name_plural = 'EsporteMuz'
-
-    # def ready(self):
-    #     import apps.esportemuz.signals
