+++ apps/esportemuz/templates/esportemuz/pages/index.html
@@ -1,0 +1,5 @@
+{% extends "esportemuz/base.html" %}
+
+{% block content %}
+<h1>Home</h1>
+{% endblock content %}
