+++ apps/esportemuz/urls.py
@@ -1,18 +1,11 @@
+from apps.esportemuz.routers import router
 from apps.esportemuz.views import *
 from django.urls import include, path
-from rest_framework.routers import DefaultRouter
-
-router = DefaultRouter()
-
-router.register(r'modalidades', ModalidadeViewSet)
-router.register(r'equipes', EquipeViewSet)
-router.register(r'tipos', TipoCampeonatoViewSet)
-router.register(r'campeonatos', CampeonatoViewSet)
-router.register(r'locais', LocalPartidaViewSet)
-router.register(r'partidas', PartidaViewSet)
-router.register(r'classificacoes', ClassificacaoViewSet)
 
 urlpatterns = [
     path('api/', include(router.urls)),
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
+    path('', IndexView.as_view(), name='index'),
+    path('equipe/', EquipeView.as_view(), name='equipe'),
+    path('campeonato/', CampeonatoView.as_view(), name='campeonato'),
 ]
