+++ apps/eventuz/urls.py
@@ -1,5 +1,5 @@
+from apps.eventuz.views import IndexView, EventosView, CadastrarEventosView
 from django.urls import path
-from .views import IndexView, EventosView, CadastrarEventosView
 
 urlpatterns = [
     path('', IndexView.as_view(), name='index'),
