+++ apps/movemuz/admin.py
@@ -49,15 +49,6 @@ class VeiculoAdmin(admin.ModelAdmin):
     foto_preview.allow_tags = True
     foto_preview.short_description = "Prévia da Foto"
 
-   
-@admin.register(Motorista)
-class MotoristaAdmin(admin.ModelAdmin):
-    list_display = ('nome', 'cpf', 'telefone', 'email', 'cnh_numero', 'data_nascimento', 'ativo')
-    search_fields = ('nome', 'cpf', 'cnh_numero')
-    list_filter = ('ativo', 'cnh_numero')
-    ordering = ('nome',)
-    list_per_page = 20
-
 @admin.register(Local)
 class LocalAdmin(admin.ModelAdmin):
     list_display = ('nome', 'cidade', 'estado')
@@ -79,14 +70,6 @@ class PassageiroAdmin(admin.ModelAdmin):
     list_display = ('nome', 'documento', 'viagem')
     search_fields = ('nome', 'documento', 'viagem__origem__nome', 'viagem__destino__nome')
     list_filter = ('viagem',)
-    ordering = ('nome',)
-
-
-@admin.register(Motorista)
-class MotoristaAdmin(admin.ModelAdmin):
-    list_display = ('nome', 'cpf', 'cnh_numero', 'ativo')
-    search_fields = ('nome', 'cpf', 'cnh_numero')
-    list_filter = ('ativo',)
     ordering = ('nome',)
