+++ apps/movemuz/urls.py
@@ -1,6 +1,5 @@
-from django.contrib import admin
+from apps.movemuz import views
 from django.urls import path
-from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
