+++ apps/muzeu/urls.py
@@ -1,8 +1,6 @@
-from django.contrib import admin
+from apps.muzeu.views import *
 from django.urls import path
-from .views import *
 
 urlpatterns = [
     path('', IndexView.as_view(), name='index'),
-    
 ]
