+++ apps/muzsaude/urls.py
@@ -1,5 +1,5 @@
+from apps.muzsaude.views import *
 from django.urls import path
-from .views import *
 
 urlpatterns = [
